refactor: normalize demo names
