+++ demo/src/damping/DampingExample.tsx
