+++ demo/src/instances-meshes-compound/InstancedMeshesCompoundExample.tsx
